changed where create song took place to inside the song form, removed it frmo the musical show. Changed styling to be consisitant

+++ Musical-app/database/migrations/2025_11_04_135426_create_songs_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('musical_id')->constrained()->onDelete('cascade');
             $table->string('title');
-            $table->integer('duration');
+            $table->decimal('duration', 5, 2);
             $table->string('composer');
             $table->timestamps();
         });
